Get table of grid resolution information

Gets a grid's resolution and cell property info as a data frame.

Usage

dggetres(dggs)

Arguments

dggs

A dggs object from dgconstruct()

Value

A data frame containing the resolution levels, number of cells, area of those cells, intercell spacing, and characteristic length scale of the cells. All values are in kilometres.

Examples

library(dggridR)
dggetres(dgconstruct(res=20))
#>    res        cells      area_km   spacing_km       cls_km
#> 1    0 1.200000e+01 5.100656e+07 7.053652e+03 8.199500e+03
#> 2    1 3.200000e+01 1.700219e+07 4.072428e+03 4.678970e+03
#> 3    2 9.200000e+01 5.667396e+06 2.351217e+03 2.691252e+03
#> 4    3 2.720000e+02 1.889132e+06 1.357476e+03 1.551868e+03
#> 5    4 8.120000e+02 6.297106e+05 7.837392e+02 8.956018e+02
#> 6    5 2.432000e+03 2.099035e+05 4.524920e+02 5.170050e+02
#> 7    6 7.292000e+03 6.996785e+04 2.612464e+02 2.984793e+02
#> 8    7 2.187200e+04 2.332262e+04 1.508307e+02 1.723245e+02
#> 9    8 6.561200e+04 7.774205e+03 8.708213e+01 9.949109e+01
#> 10   9 1.968320e+05 2.591402e+03 5.027689e+01 5.744111e+01
#> 11  10 5.904920e+05 8.638006e+02 2.902738e+01 3.316362e+01
#> 12  11 1.771472e+06 2.879335e+02 1.675896e+01 1.914702e+01
#> 13  12 5.314412e+06 9.597785e+01 9.675792e+00 1.105454e+01
#> 14  13 1.594323e+07 3.199262e+01 5.586321e+00 6.382340e+00
#> 15  14 4.782969e+07 1.066421e+01 3.225264e+00 3.684846e+00
#> 16  15 1.434891e+08 3.554735e+00 1.862107e+00 2.127447e+00
#> 17  16 4.304672e+08 1.184912e+00 1.075088e+00 1.228282e+00
#> 18  17 1.291402e+09 3.949706e-01 6.207024e-01 7.091489e-01
#> 19  18 3.874205e+09 1.316569e-01 3.583627e-01 4.094273e-01
#> 20  19 1.162261e+10 4.388562e-02 2.069008e-01 2.363830e-01
#> 21  20 3.486784e+10 1.462854e-02 1.194542e-01 1.364758e-01
#> 22  21 1.046035e+11 4.876180e-03 6.896693e-02 7.879432e-02
#> 23  22 3.138106e+11 1.625393e-03 3.981807e-02 4.549192e-02
#> 24  23 9.414318e+11 5.417977e-04 2.298898e-02 2.626477e-02
#> 25  24 2.824295e+12 1.805992e-04 1.327269e-02 1.516397e-02
#> 26  25 8.472886e+12 6.019975e-05 7.662992e-03 8.754924e-03
#> 27  26 2.541866e+13 2.006658e-05 4.424230e-03 5.054658e-03
#> 28  27 7.625597e+13 6.688861e-06 2.554331e-03 2.918308e-03
#> 29  28 2.287679e+14 2.229620e-06 1.474743e-03 1.684886e-03
#> 30  29 6.863038e+14 7.432068e-07 8.514436e-04 9.727694e-04
#> 31  30 2.058911e+15 2.477356e-07 4.915812e-04 5.616287e-04